+++ src/tui/status_view.rs
@@ -48,14 +48,15 @@ impl StatusView {
             }
             if !des.drawer.content.entries.is_empty() {
                 if matches!(des.focus, DrawerFocus::NameSelected{..}|DrawerFocus::ValueSelected{..}) {
-                    hints.push("Hit *^q* to quit, *i* to edit the selected cell");
+                    hints.push("Hit *^q* to quit, *i* to edit the selected cell, *?* for help");
                 }
                 hints.push("Hit *^q* to quit, */* to search, *n* to create a new entry");
+                hints.push("Hit *^q* to quit, *?* for help");
                 hints.push("Hit *^q* to quit, */* to search, *^h* to toggle values visibility");
                 hints.push("Hit *^q* to quit, */* to search, arrows to select a cell");
-                hints.push("Hit *^q* to quit, *tab* to edit the next cell");
+                hints.push("Hit *^q* to quit, *tab* to edit the next cell, *?* for help");
             }
-            hints.push("Hit *^q* to quit");
+            hints.push("Hit *^q* to quit, *?* for help");
         }
         let idx = (self.drawer_display_count / 3 ) % hints.len();
         self.drawer_display_count += 1;
@@ -85,10 +86,12 @@ impl View for StatusView {
                 Alignment::Unspecified,
             )?;
         } else {
-            let s = if let DrawerState::DrawerEdit(des) = &state.drawer_state {
+            let s = if state.help.is_some() {
+                "Hit *^q* to quit, *esc* to close the help"
+            } else if let DrawerState::DrawerEdit(des) = &state.drawer_state {
                 self.rotate_drawer_hint(des)
             } else {
-                "Hit *^q* to quit"
+                "Hit *^q* to quit, *?* for help"
             };
             self.hint_skin.write_composite_fill(
                 w,