+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+### next
+- help screen
+- ctrl-c to close a drawer or the help screen
+
 <a name="v0.2.0"></a>
 ### v0.2.0 - 2021-09-19
 - mouse support in inputs and for cell selection
-- sub-drawers
+- sub-drawers (and breaking change in closet format)
 
 <a name="v0.1.3"></a>
 ### v0.1.3 - 2021-09-08